requirements.txt
streamlit
yfinance
plotly
pandas
numpy
langchain
langgraph
transformers
torch
groq
python-dotenv